Everything is imported except for the menu layout

+++ LeafToAutomationImport/LeafDataModel/JobCode.cs
@@ -9,6 +9,6 @@
         public decimal rate2 { get; set; }
         public decimal otRate1 { get; set; }
         public decimal otRate2 { get; set; }
-        public decimal hrsBeforeOT { get; set; }
+        public int hrsBeforeOT { get; set; }
     }
 }
